+++ Pod/ViewController.h
@@ -1,4 +1,4 @@
-#pragma once
+	#pragma once
 #include <vector>
 #include "Setting.h"
 #include "Component.h"
@@ -23,8 +23,8 @@ public:
 	virtual void viewDidLoad();
 	virtual void viewReceiveKeyDown(KeyType _keyType);
 	virtual void viewReceiveKeyUp();
-	virtual void viewWillUpdate(int _dt);
-	virtual void viewDidUpdate(int _dt);
+	virtual void viewWillUpdate(float _dt);
+	virtual void viewDidUpdate(float _dt);
 	virtual void viewWillRender();
 	virtual void viewDidRender();
 	virtual void viewWillRelease();
